pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi-mail>=1.6.2",
    "fastapi[standard]>=0.136.0",
    "greenlet>=3.4.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.1",
    "resend>=2.29.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
